refactor: rename Dashboard components to PascalCase
Rename the nine lowercase files in app/components/ui/Dashboard/ to
PascalCase matching their exported component, so every file in the
folder follows one convention. Update all importers and normalize
their paths to the ~/ alias.
