+++ app/components/ui/Dashboard/EquipmentCard.tsx
