+++ app/components/ui/Dashboard/WorkshopCard.tsx
