+++ app/routes/dashboard/addequipment.tsx
@@ -34,8 +34,8 @@ import { getWorkshops } from "~/models/workshop.server";
 import MultiSelectField from "~/components/ui/Dashboard/MultiSelectField";
 import { useState } from "react";
 import { SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar";
-import AppSidebar from "~/components/ui/Dashboard/sidebar";
-import AdminAppSidebar from "~/components/ui/Dashboard/adminsidebar";
+import AppSidebar from "~/components/ui/Dashboard/AppSidebar";
+import AdminAppSidebar from "~/components/ui/Dashboard/AdminAppSidebar";
 import { ArrowLeft } from "lucide-react";
 import { useNavigate } from "react-router";
 import * as fs from "fs";
