+++ app/routes/dashboard/addworkshop.tsx
@@ -36,8 +36,8 @@ import {
   CalendarRange as CalendarRangeIcon,
   Check as CheckIcon,
 } from "lucide-react";
-import EquipmentBookingGrid from "~/components/ui/Dashboard/equipmentbookinggrid";
-import type { SlotsByDay } from "~/components/ui/Dashboard/equipmentbookinggrid";
+import EquipmentBookingGrid from "~/components/ui/Dashboard/EquipmentBookingGrid";
+import type { SlotsByDay } from "~/components/ui/Dashboard/EquipmentBookingGrid";
 import {
   bulkBookEquipment,
   createEquipmentSlotsForOccurrence,
@@ -61,8 +61,8 @@ import { getEquipmentVisibilityDays } from "../../models/admin.server";
 import { getUser, getRoleUser } from "../../utils/session.server";
 import { logger } from "~/logging/logger";
 import { SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar";
-import AppSidebar from "~/components/ui/Dashboard/sidebar";
-import AdminAppSidebar from "~/components/ui/Dashboard/adminsidebar";
+import AppSidebar from "~/components/ui/Dashboard/AppSidebar";
+import AdminAppSidebar from "~/components/ui/Dashboard/AdminAppSidebar";
 import { ArrowLeft } from "lucide-react";
 import { useNavigate } from "react-router";
 import { Input } from "~/components/ui/input";
