+++ app/routes/dashboard/adminreports.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { useLoaderData, redirect } from "react-router-dom";
 import { SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar";
-import AdminAppSidebar from "~/components/ui/Dashboard/adminsidebar";
+import AdminAppSidebar from "~/components/ui/Dashboard/AdminAppSidebar";
 import {
   Card,
   CardContent,
