+++ app/routes/dashboard/editmembershipplan.tsx
@@ -27,8 +27,8 @@ import {
 import { getRoleUser } from "~/utils/session.server";
 import { logger } from "~/logging/logger";
 import { SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar";
-import AppSidebar from "~/components/ui/Dashboard/sidebar";
-import AdminAppSidebar from "~/components/ui/Dashboard/adminsidebar";
+import AppSidebar from "~/components/ui/Dashboard/AppSidebar";
+import AdminAppSidebar from "~/components/ui/Dashboard/AdminAppSidebar";
 import { ArrowLeft } from "lucide-react";
 import MembershipPlanForm from "~/components/ui/Dashboard/MembershipPlanForm";
