+++ app/routes/dashboard/membershipdetails.tsx
@@ -46,9 +46,9 @@ import {
 } from "~/models/membership.server";
 import { getSavedPaymentMethod, getUserById } from "~/models/user.server";
 import { SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar";
-import AppSidebar from "~/components/ui/Dashboard/sidebar";
-import AdminAppSidebar from "~/components/ui/Dashboard/adminsidebar";
-import GuestAppSidebar from "~/components/ui/Dashboard/guestsidebar";
+import AppSidebar from "~/components/ui/Dashboard/AppSidebar";
+import AdminAppSidebar from "~/components/ui/Dashboard/AdminAppSidebar";
+import GuestAppSidebar from "~/components/ui/Dashboard/GuestAppSidebar";
 import {
   Tooltip,
   TooltipContent,
