+++ app/routes/dashboard/paymentinformation.tsx
@@ -10,9 +10,9 @@ import type { ActionFunction, LoaderFunction } from "react-router-dom";
 import { loadStripe } from "@stripe/stripe-js";
 import { Elements, CardElement, useStripe, useElements } from "@stripe/react-stripe-js";
 import { SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar";
-import AppSidebar from "../../components/ui/Dashboard/sidebar";
-import AdminAppSidebar from "../../components/ui/Dashboard/adminsidebar";
-import GuestAppSidebar from "../../components/ui/Dashboard/guestsidebar";
+import AppSidebar from "~/components/ui/Dashboard/AppSidebar";
+import AdminAppSidebar from "~/components/ui/Dashboard/AdminAppSidebar";
+import GuestAppSidebar from "~/components/ui/Dashboard/GuestAppSidebar";
 import { CountryDropdown } from "@/components/ui/country-dropdown";
 import { getUserId, getRoleUser } from "~/utils/session.server";
 import { getProfileDetails } from "~/models/profile.server";
